Image bucket

+++ src/main/java/za/co/photo_sharing/app_ws/model/response/SkillSetRest.java
@@ -12,5 +12,5 @@ public class SkillSetRest {
     private long id;
     private String skill;
     private double rating;
-    private double ratingCalc;
+    private int ratingCalc;
 }
